More work on test build script.
Added in more strucure to test-build.sh script. It now shows the time it
took to run. Also changed how the git repo is cloned, it now has a
custom build directory and custom log file.

Added build directories to .gitignore and also added them to the make
clean rule.