Added new compiler Flag: -Wextra
I decided to go ahead and add in the -Wextra flag to the compile flags. This enables things like ambigious base virtual classes, ambigious braces and whole bunch of other stuff as warnings. We might end up taking it out if it becomes a pain but I think we should try to compile our stuff pretty hardcore at first.