Added text for MAC OS X installs to Readme; Removed `sudo`
Updated README.md to have a secton for MAC OS X. I have one at work so
I decided to go ahead and start this section. I'm not sure if we should
include instructions for intalled Homebrew or not. I have it linked to
their website now.

I removed `sudo` from all the instructions. It was on almost every single
one so it was just useless clutter to me. I added a note at the top of
the depends section to clarify that we have ommited that from the
instructions and that the user should add it in as neccessary.