Adding more method definitions.
Adding the rest of the method definitions. They are currently commented
out so uncomment then and create the implemention to get going.

There are a class of operators that I have not included and those are
the ones the modify the LHS in place (+=, -=, etc). I might add the
definitions soon but at the moment they are not really on the top of my
list. I would much rather get all of the 'Properties' finished first.